+++ config/config.lua
@@ -18,6 +18,9 @@ C["Lua"] = {
     -- Media
     ["Font"] = "Pixel",                             -- default font type.
     ["Texture"] = "Blank",                          -- default texture type.
+
+    -- Layout
+    ["HealerLayout"] = true,                        -- enables healing raid layout.
 }
 
 -- UnitsFrames Size
@@ -69,7 +72,7 @@ C["Interrupts"] = {
 C["ScreenShots"] = {
     ["Enable"] = true,                              -- enables achievement/encounter screenshot plugin.
     ["BossKills"] = false,                          -- enables screenshots of successful raid encounters.
-    ["ChallendeMode"] = true,                       -- enables screenshots of successful challenge modes.
+    ["ChallendeMode"] = false,                      -- enables screenshots of successful challenge modes.
     ["Achievements"] = true,                        -- enables screenshots of earned achievements.
     ["LevelUp"] = false,                            -- enables screenshots when player level up.
     ["Messages"] = false,                           -- enables print messages when screenshot is taken or failed.